Extracted some API config into it's own file as an extension function. Might be worth combining the files in the future for vertical reasons, but I prefer this to the extremely tall Root.kt we had before. Added CallID and CallLogging KTOR plugins to simplify some of our logging configuration. These replaced the CorrelationIdPlugin, which was a bit confusing and gross IMO.
