+++ .env.example
@@ -8,6 +8,7 @@ cookie.secure=HTTPS_ONLY_BOOLEAN
 cors.url=ALLOW_HOST_VALUE
 cors.scheme=ALLOW_HOST_SCHEME
 ui.packagePath=PATH_TO_FRONTEND_PACKAGE
+logging.cid=CORRELATION_ID_KEY
 ROOT_LOG_LEVEL=TRACE
 NETTY_LOG_LEVEL=INFO
 KTOR_LOG_LEVEL=TRACE
