+++ src/main/kotlin/com/lowbudgetlcs/api/routes/v1/series/SeriesRoutes.kt
@@ -3,8 +3,6 @@ package com.lowbudgetlcs.api.routes.v1.series
 import com.lowbudgetlcs.api.dto.games.CreateGameDto
 import com.lowbudgetlcs.api.dto.games.toDto
 import com.lowbudgetlcs.api.dto.games.toNewGame
-import com.lowbudgetlcs.api.logCall
-import com.lowbudgetlcs.api.setCidContext
 import com.lowbudgetlcs.domain.series.ISeriesService
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.application.Application
@@ -21,13 +19,10 @@ private val logger: Logger = LoggerFactory.getLogger(Application::class.java)
 fun Route.seriesRoutesV1(seriesService: ISeriesService) {
     route("/series") {
         post<SeriesResources.Game> { route ->
-            call.setCidContext {
-                logCall(call)
-                val dto = call.receive<CreateGameDto>()
-                logger.debug(dto.toString())
-                val created = seriesService.createGame(dto.toNewGame(route.seriesId))
-                call.respond(HttpStatusCode.Created, created.toDto())
-            }
+            val dto = call.receive<CreateGameDto>()
+            logger.debug(dto.toString())
+            val created = seriesService.createGame(dto.toNewGame(route.seriesId))
+            call.respond(HttpStatusCode.Created, created.toDto())
         }
     }
 }
