+++ src/main/kotlin/com/lowbudgetlcs/api/routes/v1/team/TeamRoutes.kt
@@ -1,17 +1,19 @@
 package com.lowbudgetlcs.api.routes.v1.team
 
-import com.lowbudgetlcs.api.dto.teams.*
-import com.lowbudgetlcs.api.logCall
-import com.lowbudgetlcs.api.setCidContext
+import com.lowbudgetlcs.api.dto.teams.NewTeamDto
+import com.lowbudgetlcs.api.dto.teams.PatchTeamDto
+import com.lowbudgetlcs.api.dto.teams.toDto
+import com.lowbudgetlcs.api.dto.teams.toNewTeam
+import com.lowbudgetlcs.api.dto.teams.toTeamUpdate
 import com.lowbudgetlcs.domain.team.ITeamService
 import com.lowbudgetlcs.domain.team.models.toTeamId
-import io.ktor.http.*
+import io.ktor.http.HttpStatusCode
 import io.ktor.server.application.Application
-import io.ktor.server.request.*
-import io.ktor.server.resources.*
+import io.ktor.server.request.receive
+import io.ktor.server.resources.get
 import io.ktor.server.resources.patch
 import io.ktor.server.resources.post
-import io.ktor.server.response.*
+import io.ktor.server.response.respond
 import io.ktor.server.routing.Route
 import io.ktor.server.routing.route
 import org.slf4j.Logger
@@ -22,42 +24,27 @@ private val logger: Logger = LoggerFactory.getLogger(Application::class.java)
 fun Route.teamRoutesV1(teamService: ITeamService) {
     route("/team") {
         post<TeamResources> {
-            call.setCidContext {
-                logCall(call)
-                val dto = call.receive<NewTeamDto>()
-                logger.debug(dto.toString())
-                val created = teamService.createTeam(dto.toNewTeam())
-                call.respond(HttpStatusCode.Created, created.toDto())
-            }
+            val dto = call.receive<NewTeamDto>()
+            logger.debug(dto.toString())
+            val created = teamService.createTeam(dto.toNewTeam())
+            call.respond(HttpStatusCode.Created, created.toDto())
         }
         get<TeamResources> {
-            call.setCidContext {
-                logCall(call)
-                val teams = teamService.getAllTeams()
-                call.respond(teams.map { it.toDto() })
-            }
+            val teams = teamService.getAllTeams()
+            call.respond(teams.map { it.toDto() })
         }
         get<TeamResources.ById> { route ->
-            call.setCidContext {
-                logCall(call)
-                val team = teamService.getTeam(route.teamId.toTeamId())
-                call.respond(team.toDto())
-            }
+            val team = teamService.getTeam(route.teamId.toTeamId())
+            call.respond(team.toDto())
         }
         patch<TeamResources.ById> { route ->
-            call.setCidContext {
-                logCall(call)
-                val dto = call.receive<PatchTeamDto>()
-                val updated = teamService.patchTeam(route.teamId.toTeamId(), dto.toTeamUpdate())
-                call.respond(updated.toDto())
-            }
+            val dto = call.receive<PatchTeamDto>()
+            val updated = teamService.patchTeam(route.teamId.toTeamId(), dto.toTeamUpdate())
+            call.respond(updated.toDto())
         }
         get<TeamResources.ByIdPlayers> { route ->
-            call.setCidContext {
-                logCall(call)
-                val team = teamService.getTeamWithPlayers(route.teamId.toTeamId())
-                call.respond(team.toDto())
-            }
+            val team = teamService.getTeamWithPlayers(route.teamId.toTeamId())
+            call.respond(team.toDto())
         }
     }
 }
