+++ src/main/kotlin/com/lowbudgetlcs/config/AppConfig.kt
@@ -7,6 +7,7 @@ data class AppConfig(
     val database: DatabaseConfig,
     val cookie: CookieConfig,
     val cors: CorsConfig,
+    val logging: LoggingConfig,
 )
 
 data class CookieConfig(
@@ -28,4 +29,8 @@ data class DatabaseConfig(
 data class CorsConfig(
     val url: String? = null,
     val scheme: String? = null,
+)
+
+data class LoggingConfig(
+    val cid: String
 )
