+++ src/main/resources/default.properties
@@ -6,3 +6,4 @@ cookie.expiration=3600
 cookie.secure=true
 cors.url=
 cors.scheme=
+logging.cid=CID
