+++ src/main/resources/logback.xml
@@ -2,7 +2,7 @@
   <!-- Console Appender for INFO level logs -->
   <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
       <encoder>
-          <pattern>%d{yyyy-MM-dd HH:mm:ss} [%thread] [%X{c-id}] %-5level %logger{36} - %msg%n</pattern>
+          <pattern>%d{yyyy-MM-dd}T%d{HH:mm:ssZ} [%thread] [%X{CID}] %-5level %logger{36} - %msg%n</pattern>
       </encoder>
   </appender>
